dashboard-server/pyproject.toml
[project]
name = "kindle-dashboard-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "holidays>=0.89",
    "lunardate>=0.2.2",
    "matplotlib>=3.10.8",
    "requests>=2.32.5",
    "yfinance>=1.1.0",
    "playwright>=1.49.0",
    "gunicorn>=24.1.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
